Add factorizations proving Conjecture 2 in the range [43632, 80973].
Conjecture 2 says that t(n) >= n/3 for n >= 300_000.  One thing we are
interested in is how small can the lower threshold 300_000 be?

I believe the minimal threshold is exactly 43632, that is t(n) >= n/3
for all n >= 43632, while t(n) < n/3 for n = 43631.

The factorizations included in this commit witness that the conjecture
is true in the range 43632 <= n <= 80973.  Specifically, this is done
by including files such as 43632-43632-14545.txt which witnesses
t(43632) >= 14545.  All of them have been checked using
simple_check.py , which took a few seconds per file or ~17 minutes
total to run.

The factorization for 43632 is uncompressed, while the rest are
compressed with gzip.  The program simple_check.py can check the
compressed files directly.

Almost all of these factorizations are probably optimal, though a few
likely aren't due to some computational issues.

An effort was made to use as few witnesses as possible.  For example,
t(43632) >= 14545 implies that t(n) >= n/3 for n in the range [43632,
43635], so the next number we consider is the smallest missing: 43636.

Because 80973 exceeds 8*10^4, this meets up with existing results.