Add code for LP-based 'smooth factorization' heuristic
This code adds the program 'smoothfac' that implements a heuristic
for optimizing large factors of the factorial. The heuristic is based
on observations of solutions obtained by the original LP solver.

It splits the factors of N! into smooth and non-smooth factors. The
threshold for smoothness is placed at sqrt(T). Non-smooth factors
(with primes larger sqrt(T)) are added greedily. Smooth factors are
are optimized using linear programming. The resulting LP is
substantially smaller, containing only constraints for primes smaller
than sqrt(T). This allows to solve larger problem instances, while
still retaining some of the benefits of using linear programming.